backend/pyproject.toml
[project]
name = "geoassets-backend"
version = "1.0.0"
description = "GeoAssets Intelligence — Backend API"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.111.1",
    "uvicorn>=0.30.3",
    "gunicorn>=22.0.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.3.0",
    "litellm>=1.46.6",
    "boto3>=1.34.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pyyaml>=6.0.1",
    "sse-starlette>=2.0.0",
    "scipy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
unfixable = ["F401", "F841"]
extend-select = ["I", "N"]
